DEV: Add TODO about dead code from CleanUpUploads

+++ app/jobs/scheduled/clean_up_uploads.rb
@@ -36,10 +36,14 @@ module Jobs
         .with_no_non_post_relations
 
       result.find_each do |upload|
-        if upload.sha1.present?
-          encoded_sha = Base62.encode(upload.sha1.hex)
+        next if Upload.in_use_callbacks&.any? { |callback| callback.call(upload) }
 
-          next if Upload.in_use_callbacks&.any? { |callback| callback.call(upload) }
+        if upload.sha1.present?
+          # TODO: Remove this check after UploadReferences records were created
+          encoded_sha = Base62.encode(upload.sha1.hex)
+          next if ReviewableQueuedPost.pending.where("payload->>'raw' LIKE '%#{upload.sha1}%' OR payload->>'raw' LIKE '%#{encoded_sha}%'").exists?
+          next if Draft.where("data LIKE '%#{upload.sha1}%' OR data LIKE '%#{encoded_sha}%'").exists?
+          next if UserProfile.where("bio_raw LIKE '%#{upload.sha1}%' OR bio_raw LIKE '%#{encoded_sha}%'").exists?
 
           upload.destroy
         else
